refactor: DeploymentConfig
