+++ rock/sandbox/sandbox_actor.py
@@ -25,8 +25,7 @@ from rock.admin.proto.request import SandboxCommand as Command
 from rock.admin.proto.request import SandboxCreateBashSessionRequest as CreateBashSessionRequest
 from rock.admin.proto.request import SandboxReadFileRequest as ReadFileRequest
 from rock.admin.proto.request import SandboxWriteFileRequest as WriteFileRequest
-from rock.deployments.abstract import AbstractDeployment
-from rock.deployments.config import DeploymentConfig
+from rock.deployments.abstract import AbstractDeployment, DeploymentConfig
 from rock.deployments.constants import Status
 from rock.deployments.docker import DockerDeployment
 from rock.deployments.status import ServiceStatus
